connect CPU history to CPU gauge and update from cluster
also refactored a little to move some state upwards.
Notably, the (avatar) colour is in the top level, also
the renumbering from downloaded jobs to list of which job on each core
is done at the top level.

+++ cluster_queue/server.py
@@ -140,7 +140,7 @@ def get_activity():
             float(cpu_usage_meas.split(' ')[1]) for cpu_usage_meas in cpu_usage
         ]
 
-        cpu_usage = np.array(cpu_usage)
+    cpu_usage = np.array(cpu_usage, dtype=np.int64)
 
     return {
         'time':
